Added rename, and duplicate

+++ MySecondMauiApp/ViewModels/MainPageViewModel.cs
@@ -29,6 +29,15 @@
         [RelayCommand]
         async Task DeleteRockAsync(Rock? rock)
         {
+            if (rock is null)
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Deleted",
+                    "Please select a rock to delete.",
+                    "OK");
+                return;
+            }
+
             if (rockDataService.DeleteRock(rock))
             {
                 await Shell.Current.DisplayAlert(
@@ -40,8 +49,78 @@
             {
                 await Shell.Current.DisplayAlert(
                     "No Rock Deleted",
-                    "Your rock was not successfully deleted or was not selected.",
+                    "Your rock was not successfully deleted.",
                     "OK");
+            }
+        }
+
+        [RelayCommand]
+        async Task DuplicateRockAsync(Rock? rock)
+        {
+            if (rock is null)
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Duplicated",
+                    "Please select a rock to duplicate.",
+                    "OK");
+                return;
+            }
+
+            if (rockDataService.DuplicateRock(rock))
+            {
+                await Shell.Current.DisplayAlert(
+                    "Rock Duplicated",
+                    "Your rock has been duplicated successfully!",
+                    "OK");
+            }
+            else
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Duplicated",
+                    "Your rock was not successfully duplicated.",
+                    "OK");
+            }
+        }
+        [RelayCommand]
+        async Task RenameRock(Rock? rock)
+        {
+            if (rock is null)
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Renamed",
+                    "Please select a rock to rename.",
+                    "OK");
+                return;
+            }
+
+            string? name = await Shell.Current.DisplayPromptAsync("Rename", "Enter new Rock name:");
+
+            if (rockDataService.ChangeName(rock, name))
+            {
+                await Shell.Current.DisplayAlert(
+                    "Rock Renamed",
+                    "Your rock has been renamed successfully!",
+                    "OK");
+            }
+            else
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Renamed",
+                    "Your rock was not successfully renamed.",
+                    "OK");
+            }
+        }
+
+        [RelayCommand]
+        async Task DownloadRock(Rock? rock)
+        {
+            if (rock is null)
+            {
+                await Shell.Current.DisplayAlert(
+                    "No Rock Downloaded",
+                    "Please select a rock to download.",
+                    "OK");
+                return;
             }
         }
     }
